Feat/application

+++ src/main/java/clubhouse/clubhouse/domain/club/dto/ClubMembersForm.java
@@ -5,11 +5,11 @@ import lombok.Data;
 @Data
 public class ClubMembersForm {
     private String name;
-    private Long age;
+    private int age;
     private String univ;
     private String major;
 
-    public ClubMembersForm(String name, Long age, String univ, String major) {
+    public ClubMembersForm(String name, int age, String univ, String major) {
         this.name = name;
         this.age = age;
         this.univ = univ;
